Add files via upload
This is my very first foray into coding.
It is a throwback to what I recall from the 80's in the arcades.
This was a fantastic bit of code to start with, teaching me invaluable lessons.  As it was my first code, every piece of it was a learning process, and since it is literally coded top down, it has a lot to be improved upon, but I had to learn what "import" is for, then what to import to hear sound, how to access sound files, images, how to put them on the screen then manipulate them.  One of the hardest parts of this was learning how to restart the game...  The last bit of code I added, and the most difficult :)

Place all files in the same directory
The code has components that are Microsoft Windows specific (the parts I used to determine screen dimensions to pass to the game "ctypes").

Python 3.7 or newer
You will need the following modules:  pygame, time, random, statistics, ctypes, os and sys

Enjoy!  and be kind :)